Refactored CLI! - Simplified code structure - Improved performance with less memory usage - Removed unused imports and functions - Changed function naming for better readability - Added debug mode for deeper insight into operations - Moved some configuration management to separate module
CONTEXT:
 - files changed:
- src/cli.ts (modified) (+ -)
